Translate confli.f90 to C (cs_boundary_condition_set_coeffs)
Edited by Y. Fournier.

Includes many C binding improvements, and various style fixes.

+++ src/base/cavitation.f90
@@ -201,7 +201,7 @@ contains
 
     ! Arguments
 
-    double precision pressure(ncelet), voidf(ncelet)
+    real(c_double) pressure(ncelet), voidf(ncelet)
 
     ! Local variables
